Move RestartManager to common core project. Change to return IEnumerable. Fix possible issues with size of list changing.
