+++ src/R/Components/Impl/Microsoft.R.Components.csproj
@@ -132,7 +132,6 @@
     <Compile Include="InteractiveWorkflow\IRInteractiveWorkflow.cs" />
     <Compile Include="InteractiveWorkflow\IRInteractiveWorkflowOperations.cs" />
     <Compile Include="InteractiveWorkflow\IRInteractiveWorkflowProvider.cs" />
-    <Compile Include="PackageManager\Implementation\RestartManager.cs" />
     <Compile Include="PackageManager\Implementation\SessionPool.cs" />
     <Compile Include="PackageManager\Implementation\ViewModel\RPackageSourceViewModel.cs" />
     <Compile Include="PackageManager\Implementation\ViewModel\RPackageViewModel.cs" />
